+++ R/shiny_template.R
@@ -2,11 +2,11 @@
 
 #' @description Run a shiny template with option to download.
 #'
-#' @param template template name. Available templates are "template1" with graph and table tabs, and "template2" also with a map tab. Defaults to "template1".
+#' @param template template name. Available templates are 1 with graph and table tabs, and 2 also with a map tab. Defaults to "template1".
 #' @param ... passed to \code{shiny::runApp}
 #'
 #' @export
-shiny_template <- function(template = "template1", ...) {
+shiny_template <- function(template = 1, ...) {
   suppressWarnings({
     
     templatedir <- system.file("shiny", package = "simplevis")
@@ -15,8 +15,7 @@ shiny_template <- function(template = "template1", ...) {
     
     if (!template %in% templates)
       stop(sprintf(
-        "No such template.  Available templates, and corresponding examples, are:\n%s",
-        paste(templates, collapse = "\n")
+        "No such template.  Available templates are 1 and 2"
       ))
     
     shiny::runApp(sprintf("%s/%s", templatedir, template), ...)