+++ NEWS.md
@@ -1,10 +1,14 @@
+# simplevis 5.0.6
+
+* Renamed run_template to shiny_template with template names 1 and 2.
+* Made gg_theme_map an internal function.
+
 # simplevis 5.0.5 
 
 * Breaking: Removed *_gridlines_minor arguments.
 * Breaking: Removed font_family, font_size_title and font_size_body arguments.
 * Incorporated new theme method into gg functions.
 * Updated shiny templates for new methods.
-* Renamed run_template to shiny_template.
 
 # simplevis 5.0.4